backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-hcl2
networkx